RD-3949: Extracted validation, form fields into separate files
