+++ app/components/ContactDetailsModal/ContactDetailsForm/ContactDetailsForm.tsx
@@ -1,82 +1,8 @@
 import React from 'react';
 import { useErrors, useInputs } from '../../../utils/hooks';
 import { Modal, Form, UnsafelyTypedFormField, ApproveButton } from '../../basic';
-
-enum FormFieldType {
-    Text,
-    Checkbox
-}
-
-interface FormFieldValidation {
-    regexp: RegExp;
-    errorMessage: string;
-}
-
-interface FormField {
-    name: string;
-    label: string;
-    isRequired?: boolean;
-    type: FormFieldType;
-    validation?: FormFieldValidation;
-}
-
-const formFields: FormField[] = [
-    {
-        name: 'firstName',
-        label: 'First name',
-        isRequired: true,
-        type: FormFieldType.Text,
-        validation: {
-            errorMessage: 'Should be between 2 and 20 characters',
-            regexp: /^(.){2,20}$/
-        }
-    },
-    {
-        name: 'lastName',
-        label: 'Last name',
-        isRequired: true,
-        type: FormFieldType.Text,
-        validation: {
-            errorMessage: 'Should be between 2 and 20 characters',
-            regexp: /^(.){2,20}$/
-        }
-    },
-    {
-        name: 'email',
-        label: 'Email address',
-        isRequired: true,
-        type: FormFieldType.Text,
-        validation: {
-            errorMessage: 'Should be a valid email address',
-            regexp: /^\S+@\S+\.\S+$/
-        }
-    },
-    {
-        name: 'phone',
-        label: 'Phone number',
-        isRequired: true,
-        type: FormFieldType.Text,
-        validation: {
-            errorMessage: 'Should be between 4 and 20 digits',
-            regexp: /^(\d){4,20}$/
-        }
-    },
-    {
-        name: 'isEULA',
-        label:
-            'By registering for the Cloudify Hosted Service you agree to the Hosted Trial End User License Agreement.',
-        isRequired: true,
-        type: FormFieldType.Checkbox
-    },
-    {
-        name: 'isSendServicesDetails',
-        label:
-            'Cloudify uses the information provided to send you your service details. For more information see our Privacy Policy.',
-        type: FormFieldType.Checkbox
-    }
-];
-
-const requiredFormFields = formFields.filter(formField => formField.isRequired);
+import type { FormField } from './form-fields';
+import { FormFieldType, formFields, requiredFormFields } from './form-fields';
 
 type FormFieldValue = string | boolean;
